@@ -15,7 +15,7 @@
     "@catppuccin/daisyui": "^1.1.2",
     "@sveltejs/adapter-vercel": "^5.4.4",
     "@sveltejs/kit": "^2.5.28",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/node": "^22.5.5",
     "autoprefixer": "^10.4.20",
     "daisyui": "^4.12.10",
@@ -27,7 +27,7 @@
     "prettier": "^3.3.3",
     "prettier-plugin-svelte": "^3.2.6",
     "prettier-plugin-tailwindcss": "^0.6.6",
-    "svelte": "^4.2.19",
+    "svelte": "^5.0.0",
     "svelte-check": "^4.0.2",
     "sveltekit-superforms": "^2.16.1",
     "tailwindcss": "^3.4.12",
