@@ -2,9 +2,13 @@
   import { invalidate } from "$app/navigation";
   import { Trash } from "lucide-svelte";
 
-  export let id: string;
-  let loading = false;
-  let modal: HTMLDialogElement;
+  interface Props {
+    id: string;
+  }
+
+  let { id }: Props = $props();
+  let loading = $state(false);
+  let modal: HTMLDialogElement = $state();
 
   async function deleteImage() {
     modal.close();
@@ -26,8 +30,8 @@
     <h3 class="text-center text-lg font-bold">Delete {id}?</h3>
 
     <div class="mt-4 flex w-full justify-center gap-2">
-      <button class="btn btn-error" on:click={deleteImage}>Delete</button>
-      <button class="btn" on:click={() => modal.close()}>Cancel</button>
+      <button class="btn btn-error" onclick={deleteImage}>Delete</button>
+      <button class="btn" onclick={() => modal.close()}>Cancel</button>
     </div>
   </div>
   <form method="dialog" class="modal-backdrop backdrop-blur-lg">
@@ -38,10 +42,10 @@
 <button
   class="btn btn-ghost tooltip tooltip-top tooltip-error text-error"
   data-tip="delete"
-  on:click={() => modal.show()}
+  onclick={() => modal.show()}
 >
   {#if loading}
-    <span class="loading loading-spinner loading-xs text-error" />
+    <span class="loading loading-spinner loading-xs text-error"></span>
   {:else}
     <Trash size={16} strokeWidth={2.5} />
   {/if}
