@@ -5,6 +5,11 @@
   import { onMount } from "svelte";
   import { Toaster } from "svelte-french-toast";
   import "../app.css";
+  interface Props {
+    children?: import('svelte').Snippet;
+  }
+
+  let { children }: Props = $props();
 
   onMount(() => {
     getLocalAuth();
@@ -25,4 +30,4 @@
 
 <Navbar />
 
-<slot />
+{@render children?.()}
